Organize dataset base classes.
