+++ streaming/base/__init__.py
@@ -3,9 +3,9 @@
 
 """MosaicML Streaming Datasets for cloud-native model training."""
 
-from streaming.base.dataset import Dataset
+from streaming.base.dataset import (Dataset, LocalIterableDataset, LocalMapDataset,
+                                    LocalResumableDataset)
 from streaming.base.format import CSVWriter, JSONWriter, MDSWriter, TSVWriter, XSVWriter
-from streaming.base.local import LocalIterableDataset, LocalMapDataset, LocalResumableDataset
 
 __all__ = [
     'Dataset', 'CSVWriter', 'JSONWriter', 'LocalIterableDataset', 'LocalMapDataset',
